feat: 상품 수정 중 오류 처리 구현 #9

+++ OpenMarket/OpenMarket/Network/API/APIModel/UpdateProductRequestModel/UpdateProductRequestModel.swift
@@ -12,9 +12,9 @@ struct UpdateProductRequestModel: Encodable {
     let name: String?
     let descriptions: String?
     let thumbnailID: Int?
-    let price: Decimal?
+    let price: Double?
     let currency: Currency?
-    let discountedPrice: Int?
+    let discountedPrice: Double?
     let secret: String
     
     enum CodingKeys: String, CodingKey {
